+++ database/migrations/2020_03_18_165210_alter_table_action_instance_fields_rename_event_field_to_action_value.php
@@ -1,0 +1,45 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AlterTableActionInstanceFieldsRenameEventFieldToActionValue extends Migration
+{
+    /**
+     * Run the migrations.
+     * 
+     * @return void
+     */
+    public function up()
+    {
+        $actionInstanceFields = \BristolSU\Support\Action\ActionInstanceField::all();
+        foreach($actionInstanceFields as $actionInstanceField) {
+            $actionInstanceField->event_field = sprintf('{{event:%s}}', $actionInstanceField->event_field);
+            $actionInstanceField->save();
+        }
+        
+        Schema::table('action_instance_fields', function (Blueprint $table) {
+            $table->renameColumn('event_field', 'action_value');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('action_instance_fields', function (Blueprint $table) {
+            $table->renameColumn('action_value', 'event_field');
+        });
+
+        $actionInstanceFields = \BristolSU\Support\Action\ActionInstanceField::all();
+        foreach($actionInstanceFields as $actionInstanceField) {
+            $actionInstanceField->event_field = substr(substr($actionInstanceField->event_field, 8), 0, -2);
+            $actionInstanceField->save();
+        }
+        
+    }
+}